Fix Propel driver.
This commit fixes/extends Propel driver by adding Listener class with
events from GlorpenPropelBundle.
