+++ Doctrine/Listener.php
@@ -3,92 +3,13 @@
 namespace FOS\ElasticaBundle\Doctrine;
 
 use Doctrine\Common\Persistence\Event\LifecycleEventArgs;
-use FOS\ElasticaBundle\Persister\ObjectPersister;
-use FOS\ElasticaBundle\Persister\ObjectPersisterInterface;
-use FOS\ElasticaBundle\Provider\IndexableInterface;
-use Psr\Log\LoggerInterface;
-use Symfony\Component\PropertyAccess\PropertyAccess;
-use Symfony\Component\PropertyAccess\PropertyAccessorInterface;
 
 /**
  * Automatically update ElasticSearch based on changes to the Doctrine source
  * data. One listener is generated for each Doctrine entity / ElasticSearch type.
  */
-class Listener
+class Listener extends AbstractListenerBase
 {
-    /**
-     * Object persister.
-     *
-     * @var ObjectPersisterInterface
-     */
-    protected $objectPersister;
-
-    /**
-     * Configuration for the listener.
-     *
-     * @var array
-     */
-    private $config;
-
-    /**
-     * Objects scheduled for insertion.
-     *
-     * @var array
-     */
-    public $scheduledForInsertion = array();
-
-    /**
-     * Objects scheduled to be updated or removed.
-     *
-     * @var array
-     */
-    public $scheduledForUpdate = array();
-
-    /**
-     * IDs of objects scheduled for removal.
-     *
-     * @var array
-     */
-    public $scheduledForDeletion = array();
-
-    /**
-     * PropertyAccessor instance.
-     *
-     * @var PropertyAccessorInterface
-     */
-    protected $propertyAccessor;
-
-    /**
-     * @var IndexableInterface
-     */
-    private $indexable;
-
-    /**
-     * Constructor.
-     *
-     * @param ObjectPersisterInterface $objectPersister
-     * @param IndexableInterface       $indexable
-     * @param array                    $config
-     * @param LoggerInterface          $logger
-     */
-    public function __construct(
-        ObjectPersisterInterface $objectPersister,
-        IndexableInterface $indexable,
-        array $config = array(),
-        LoggerInterface $logger = null
-    ) {
-        $this->config = array_merge(array(
-            'identifier' => 'id',
-        ), $config);
-        $this->indexable = $indexable;
-        $this->objectPersister = $objectPersister;
-        $this->propertyAccessor = PropertyAccess::createPropertyAccessor();
-
-        if ($logger && $this->objectPersister instanceof ObjectPersister) {
-            $this->objectPersister->setLogger($logger);
-        }
-    }
-
     /**
      * Looks for new objects that should be indexed.
      *
@@ -96,11 +17,7 @@ class Listener
      */
     public function postPersist(LifecycleEventArgs $eventArgs)
     {
-        $entity = $eventArgs->getObject();
-
-        if ($this->objectPersister->handlesObject($entity) && $this->isObjectIndexable($entity)) {
-            $this->scheduledForInsertion[] = $entity;
-        }
+        $this->doPostPersist($eventArgs->getObject());
     }
 
     /**
@@ -110,16 +27,7 @@ class Listener
      */
     public function postUpdate(LifecycleEventArgs $eventArgs)
     {
-        $entity = $eventArgs->getObject();
-
-        if ($this->objectPersister->handlesObject($entity)) {
-            if ($this->isObjectIndexable($entity)) {
-                $this->scheduledForUpdate[] = $entity;
-            } else {
-                // Delete if no longer indexable
-                $this->scheduleForDeletion($entity);
-            }
-        }
+        $this->doPostUpdate($eventArgs->getObject());
     }
 
     /**
@@ -130,31 +38,7 @@ class Listener
      */
     public function preRemove(LifecycleEventArgs $eventArgs)
     {
-        $entity = $eventArgs->getObject();
-
-        if ($this->objectPersister->handlesObject($entity)) {
-            $this->scheduleForDeletion($entity);
-        }
-    }
-
-    /**
-     * Persist scheduled objects to ElasticSearch
-     * After persisting, clear the scheduled queue to prevent multiple data updates when using multiple flush calls.
-     */
-    private function persistScheduled()
-    {
-        if (count($this->scheduledForInsertion)) {
-            $this->objectPersister->insertMany($this->scheduledForInsertion);
-            $this->scheduledForInsertion = array();
-        }
-        if (count($this->scheduledForUpdate)) {
-            $this->objectPersister->replaceMany($this->scheduledForUpdate);
-            $this->scheduledForUpdate = array();
-        }
-        if (count($this->scheduledForDeletion)) {
-            $this->objectPersister->deleteManyByIdentifiers($this->scheduledForDeletion);
-            $this->scheduledForDeletion = array();
-        }
+        $this->doPreRemove($eventArgs->getObject());
     }
 
     /**
@@ -170,7 +54,7 @@ class Listener
      */
     public function preFlush()
     {
-        $this->persistScheduled();
+        $this->doPostFlush();
     }
 
     /**
@@ -179,34 +63,6 @@ class Listener
      */
     public function postFlush()
     {
-        $this->persistScheduled();
-    }
-
-    /**
-     * Record the specified identifier to delete. Do not need to entire object.
-     *
-     * @param object $object
-     */
-    private function scheduleForDeletion($object)
-    {
-        if ($identifierValue = $this->propertyAccessor->getValue($object, $this->config['identifier'])) {
-            $this->scheduledForDeletion[] = $identifierValue;
-        }
-    }
-
-    /**
-     * Checks if the object is indexable or not.
-     *
-     * @param object $object
-     *
-     * @return bool
-     */
-    private function isObjectIndexable($object)
-    {
-        return $this->indexable->isObjectIndexable(
-            $this->config['indexName'],
-            $this->config['typeName'],
-            $object
-        );
+        $this->doPostFlush();
     }
 }
