+++ Resources/config/propel.xml
@@ -3,23 +3,37 @@
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
     xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
 
+	<parameters>
+        <parameter key="fos_elastica.provider.prototype.propel.class">FOS\ElasticaBundle\Propel\Provider</parameter>
+        <parameter key="fos_elastica.listener.prototype.propel.class">FOS\ElasticaBundle\Propel\Listener</parameter>
+        <parameter key="fos_elastica.elastica_to_model_transformer.prototype.propel.class">FOS\ElasticaBundle\Propel\ElasticaToModelTransformer</parameter>
+        <parameter key="fos_elastica.manager.propel.class">FOS\ElasticaBundle\Doctrine\RepositoryManager</parameter>
+    </parameters>
+
     <services>
-        <service id="fos_elastica.provider.prototype.propel" class="FOS\ElasticaBundle\Propel\Provider" public="true" abstract="true">
+        <service id="fos_elastica.provider.prototype.propel" class="%fos_elastica.provider.prototype.propel.class%" public="true" abstract="true">
             <argument /> <!-- object persister -->
             <argument type="service" id="fos_elastica.indexable" />
             <argument /> <!-- model -->
             <argument type="collection" /> <!-- options -->
         </service>
 
-        <service id="fos_elastica.elastica_to_model_transformer.prototype.propel" class="FOS\ElasticaBundle\Propel\ElasticaToModelTransformer" public="false">
+        <service id="fos_elastica.elastica_to_model_transformer.prototype.propel" class="%fos_elastica.elastica_to_model_transformer.prototype.propel.class%" public="false">
             <argument /> <!-- model -->
             <argument type="collection" /> <!-- options -->
             <call method="setPropertyAccessor">
                 <argument type="service" id="fos_elastica.property_accessor" />
             </call>
         </service>
+        
+        <service id="fos_elastica.listener.prototype.propel" class="%fos_elastica.listener.prototype.propel.class%" public="true" abstract="true">
+            <argument /> <!-- object persister -->
+            <argument type="service" id="fos_elastica.indexable" />
+            <argument type="collection" /> <!-- configuration -->
+            <argument>null</argument> <!-- logger -->
+        </service>
 
-        <service id="fos_elastica.manager.propel" class="FOS\ElasticaBundle\Doctrine\RepositoryManager">
+        <service id="fos_elastica.manager.propel" class="%fos_elastica.manager.propel.class%">
             <argument type="service" id="annotation_reader"/>
         </service>
     </services>
